# swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "르방이들 API 문서",
    "description": "르방이들 프로젝트 API 명세서 입니다.",
    "version": "1.0.0",
    "license": "MIT"
  },
  "host": "http://levains-lb-2013408822.ap-northeast-2.elb.amazonaws.com",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/users": {
      "get": {
        "summary": "홈 화면",
        "description": "홈 화면 API",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "category_items": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "user_id": {
                        "type": "number",
                        "description": "유저 PK"
                      },
                      "username": {
                        "type": "string",
                        "description": "닉네임"
                      },
                      "kakao_talk_chatting_url": {
                        "type": "string",
                        "description": "채팅 URL"
                      },
                      "item_id": {
                        "type": "number",
                        "description": "아이템 PK"
                      },
                      "item_type": {
                        "type": "string",
                        "description": "상품이 내꺼인지 상대꺼 인지"
                      },
                      "category": {
                        "type": "string",
                        "description": "상품 카테고리"
                      },
                      "trade_status": {
                        "type": "string",
                        "description": "상품 상"
                      },
                      "name": {
                        "type": "string",
                        "description": "상품 이름"
                      },
                      "img_name": {
                        "type": "string",
                        "description": "이미지 이름"
                      },
                      "img_url": {
                        "type": "string",
                        "description": "이미지 URL"
                      }
                    }
                  }
                },
                "experience_items": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "user_id": {
                        "type": "number",
                        "description": "유저 PK"
                      },
                      "username": {
                        "type": "string",
                        "description": "닉네임"
                      },
                      "kakao_talk_chatting_url": {
                        "type": "string",
                        "description": "채팅 URL"
                      },
                      "item_id": {
                        "type": "number",
                        "description": "아이템 PK"
                      },
                      "item_type": {
                        "type": "string",
                        "description": "상품이 내꺼인지 상대꺼 인지"
                      },
                      "category": {
                        "type": "string",
                        "description": "상품 카테고리"
                      },
                      "trade_status": {
                        "type": "string",
                        "description": "상품 상"
                      },
                      "name": {
                        "type": "string",
                        "description": "상품 이름"
                      },
                      "img_name": {
                        "type": "string",
                        "description": "이미지 이름"
                      },
                      "img_url": {
                        "type": "string",
                        "description": "이미지 URL"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "요청 파라미터가 부족합니다",
            "schema": {
              "properties": {
                "message": {
                  "type": "string",
                  "example": "요청 파라미터가 부족합니다"
                }
              }
            }
          },
          "404": {
            "description": "페이지를 찾을 수 없습니다",
            "schema": {
              "properties": {
                "message": {
                  "type": "string",
                  "example": "페이지를 찾을 수 없습니다"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류입니다",
            "schema": {
              "properties": {
                "message": {
                  "type": "string",
                  "example": "서버 내부 오류입니다"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/sign-in": {
      "post": {
        "summary": "로그인",
        "description": "사용자 로그인 API",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "description": "사용자 아이디"
                },
                "kakao_talk_chatting_url": {
                  "type": "string",
                  "description": "사용자 아이디"
                },
                "role": {
                  "type": "string",
                  "description": "사용자 아이디"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "properties": {
                "access_token": {
                  "type": "string",
                  "description": "사용자 인가 액세스 토큰"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "schema": {
              "properties": {
                "access_token": {
                  "type": "string",
                  "description": "사용자 인가 액세스 토큰"
                }
              }
            }
          },
          "400": {
            "description": "요청 파라미터가 부족합니다",
            "schema": {
              "properties": {
                "message": {
                  "type": "string",
                  "example": "요청 파라미터가 부족합니다"
                }
              }
            }
          },
          "404": {
            "description": "페이지를 찾을 수 없습니다",
            "schema": {
              "properties": {
                "message": {
                  "type": "string",
                  "example": "페이지를 찾을 수 없습니다"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류입니다",
            "schema": {
              "properties": {
                "message": {
                  "type": "string",
                  "example": "서버 내부 오류입니다"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/sign-in/address": {
      "post": {
        "summary": "사용자 주소 입력",
        "description": "사용자 주소 입력 API",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "address": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "latitude": {
                        "type": "number",
                        "description": "경도",
                        "example": 123.123
                      },
                      "longitude": {
                        "type": "number",
                        "description": "위도",
                        "example": 234.234
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "요청 파라미터가 부족합니다",
            "schema": {
              "properties": {
                "message": {
                  "type": "string",
                  "example": "요청 파라미터가 부족합니다"
                }
              }
            }
          },
          "404": {
            "description": "페이지를 찾을 수 없습니다",
            "schema": {
              "properties": {
                "message": {
                  "type": "string",
                  "example": "페이지를 찾을 수 없습니다"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류입니다",
            "schema": {
              "properties": {
                "message": {
                  "type": "string",
                  "example": "서버 내부 오류입니다"
                }
              }
            }
          }
        }
      }
    }
  }
}